+++ packages/db/lib/load-config.mjs
@@ -1,54 +1,15 @@
-import parseArgs from 'minimist'
-import { access } from 'fs/promises'
+import { loadConfig } from '@platformatic/service'
 import ConfigManager from './config.js'
-import deepmerge from '@fastify/deepmerge'
-import { findConfigFile } from './utils.js'
 
-async function loadConfig (minimistConfig, _args, configOpts = {}) {
-  const args = parseArgs(_args, deepmerge({ all: true })({
-    string: ['allow-env'],
-    default: {
-      allowEnv: '' // The default is set in ConfigManager
-    },
-    alias: {
-      v: 'version',
-      c: 'config',
-      allowEnv: ['allow-env', 'E']
-    }
-  }, minimistConfig))
-  try {
-    if (!args.config) {
-      args.config = await findConfigFile(process.cwd())
-    }
-    await access(args.config)
-  } catch (err) {
-    console.error('Missing config file')
-    process.exit(1)
-  }
+const ourConfigFiles = [
+  'platformatic.db.json',
+  'platformatic.db.json5',
+  'platformatic.db.yaml',
+  'platformatic.db.yml',
+  'platformatic.db.toml',
+  'platformatic.db.tml'
+]
 
-  const configManager = new ConfigManager({
-    source: args.config,
-    envWhitelist: [...args.allowEnv.split(',')],
-    ...configOpts
-  })
-
-  const parsingResult = await configManager.parse()
-  if (!parsingResult) {
-    printConfigValidationErrors(configManager)
-    process.exit(1)
-  }
-
-  return { configManager, args }
+export default async function (a, b, c) {
+  return loadConfig(a, b, c, ConfigManager, ourConfigFiles)
 }
-
-function printConfigValidationErrors (configManager) {
-  const tabularData = configManager.validationErrors.map((err) => {
-    return {
-      path: err.path,
-      message: err.message
-    }
-  })
-  console.table(tabularData, ['path', 'message'])
-}
-
-export default loadConfig
