+++ packages/db/package.json
@@ -39,15 +39,14 @@
     "@fastify/basic-auth": "^4.0.0",
     "@fastify/cors": "^8.0.0",
     "@fastify/deepmerge": "^1.1.0",
-    "@fastify/restartable": "^1.3.1",
     "@fastify/static": "^6.5.0",
     "@fastify/swagger": "^8.0.0",
-    "@fastify/under-pressure": "^8.0.0",
     "@platformatic/config": "workspace:*",
     "@platformatic/db-authorization": "workspace:*",
     "@platformatic/db-core": "workspace:*",
     "@platformatic/db-dashboard": "workspace:*",
     "@platformatic/utils": "workspace:*",
+    "@platformatic/service": "workspace:*",
     "close-with-grace": "^1.1.0",
     "commist": "^3.1.2",
     "desm": "^1.2.0",
@@ -59,7 +58,6 @@
     "fastify-metrics": "^10.0.0",
     "fastify-plugin": "^4.1.0",
     "fastify-print-routes": "^2.0.0",
-    "fastify-sandbox": "^0.8.0",
     "graphql": "^16.6.0",
     "help-me": "^4.1.0",
     "minimatch": "^5.1.0",
