+++ packages/service/fixtures/hello/no-server-logger.json
