+++ packages/service/index.js
@@ -6,7 +6,8 @@ const underPressure = require('@fastify/under-pressure')
 const { schema } = require('./lib/schema')
 const ConfigManager = require('./lib/config.js')
 const { addLoggerToTheConfig, getJSPluginPath } = require('./lib/utils')
-const { deepmerge, isKeyEnabled } = require('@platformatic/utils')
+const loadConfig = require('./lib/load-config')
+const { isKeyEnabled, deepmerge } = require('@platformatic/utils')
 
 function createServerConfig (config) {
   // convert the config file to a new structure
@@ -16,48 +17,56 @@ function createServerConfig (config) {
   return serverConfig
 }
 
-async function platformaticService (app, opts) {
-  // Metrics plugin
-  // if (isKeyEnabled('metrics', opts)) {
-  //  app.register(require('./lib/metrics-plugin'), opts.metrics)
-  // }
-
-  let pluginOptions = opts.plugin
-
-  app.decorate('platformatic', {})
-
-  /* c8 ignore next 4 */
-  if (pluginOptions.typescript !== undefined) {
-    const pluginPath = getJSPluginPath(pluginOptions.path, pluginOptions.typescript.outDir)
-    pluginOptions = { ...pluginOptions, path: pluginPath }
+async function platformaticService (app, opts, toLoad = []) {
+  if (isKeyEnabled('metrics', opts)) {
+    app.register(require('./lib/metrics-plugin'), opts.metrics)
   }
 
-  app.log.debug({ plugin: opts.plugin }, 'loading plugin')
+  if (Array.isArray(toLoad)) {
+    for (const plugin of toLoad) {
+      await app.register(plugin, opts)
+    }
+  }
 
-  // if not defined, we defaults to true (which can happen only if config is set programmatically,
-  // that's why we ignore the coverage of the `undefined` case, which cannot be covered in cli tests)
-  /* c8 ignore next */
-  const hotReload = opts.plugin.watchOptions?.hotReload !== false
-  const isWatchEnabled = opts.plugin.watch !== false
+  if (!app.hasDecorator('platformatic')) {
+    app.decorate('platformatic', {})
+  }
 
-  if (isWatchEnabled && hotReload) {
-    await app.register(sandbox, {
-      ...pluginOptions,
-      customizeGlobalThis (_globalThis) {
-      // Taken from https://github.com/nodejs/undici/blob/fa9fd9066569b6357acacffb806aa804b688c9d8/lib/global.js#L5
-        const globalDispatcher = Symbol.for('undici.globalDispatcher.1')
-        const dispatcher = globalThis[globalDispatcher]
-        /* istanbul ignore else */
-        if (dispatcher) {
-          _globalThis[globalDispatcher] = dispatcher
+  if (opts.plugin) {
+    let pluginOptions = opts.plugin
+    /* c8 ignore next 4 */
+    if (pluginOptions.typescript !== undefined) {
+      const pluginPath = getJSPluginPath(pluginOptions.path, pluginOptions.typescript.outDir)
+      pluginOptions = { ...pluginOptions, path: pluginPath }
+    }
+
+    app.log.debug({ plugin: opts.plugin }, 'loading plugin')
+
+    // if not defined, we defaults to true (which can happen only if config is set programmatically,
+    // that's why we ignore the coverage of the `undefined` case, which cannot be covered in cli tests)
+    /* c8 ignore next */
+    const hotReload = opts.plugin.watchOptions?.hotReload !== false
+    const isWatchEnabled = opts.plugin.watch !== false
+
+    if (isWatchEnabled && hotReload) {
+      await app.register(sandbox, {
+        ...pluginOptions,
+        customizeGlobalThis (_globalThis) {
+        // Taken from https://github.com/nodejs/undici/blob/fa9fd9066569b6357acacffb806aa804b688c9d8/lib/global.js#L5
+          const globalDispatcher = Symbol.for('undici.globalDispatcher.1')
+          const dispatcher = globalThis[globalDispatcher]
+          /* istanbul ignore else */
+          if (dispatcher) {
+            _globalThis[globalDispatcher] = dispatcher
+          }
         }
-      }
-    })
-  // c8 fails in reporting the coverage of this else branch, so we ignore it
-  /* c8 ignore next 4 */
-  } else {
-    const plugin = await import(`file://${pluginOptions.path}`)
-    await app.register(plugin, pluginOptions.options)
+      })
+    // c8 fails in reporting the coverage of this else branch, so we ignore it
+    /* c8 ignore next 4 */
+    } else {
+      const plugin = await import(`file://${pluginOptions.path}`)
+      await app.register(plugin, pluginOptions.options)
+    }
   }
 
   // Enable CORS
@@ -67,15 +76,15 @@ async function platformaticService (app, opts) {
   if (isKeyEnabled('healthCheck', opts)) {
     app.register(underPressure, {
       exposeStatusRoute: '/status',
-      healthCheckInterval: opts.healthCheck.interval !== undefined ? opts.healthCheck.interval : 5000
+      healthCheckInterval: opts.healthCheck.interval !== undefined ? opts.healthCheck.interval : 5000,
+      healthCheck: opts.healthCheck.fn
     })
   }
-  await app
 }
 
 platformaticService[Symbol.for('skip-override')] = true
 
-async function buildServer (options) {
+async function buildServer (options, app = platformaticService) {
   if (!options.configManager) {
     // instantiate a new config manager from current options
     const cm = new ConfigManager({
@@ -89,7 +98,7 @@ async function buildServer (options) {
   const serverConfig = createServerConfig(options)
 
   serverConfig.originalConfig = options
-  serverConfig.app = platformaticService
+  serverConfig.app = app
   const handler = await start(serverConfig)
 
   Object.defineProperty(handler, 'url', {
@@ -112,7 +121,7 @@ async function buildServer (options) {
     /* c8 ignore start */
     if (opts) {
       opts = createServerConfig(opts)
-      opts.app = platformaticService
+      opts.app = app
 
       const fileWatcher = handler.app.platformatic.fileWatcher
       const configManager = handler.app.platformatic.configManager
@@ -137,6 +146,8 @@ async function buildServer (options) {
 }
 
 module.exports.buildServer = buildServer
-module.exports.schema = schema
+module.exports.schema = require('./lib/schema')
 module.exports.createServerConfig = createServerConfig
 module.exports.platformaticService = platformaticService
+module.exports.addLoggerToTheConfig = addLoggerToTheConfig
+module.exports.loadConfig = loadConfig
