+++ packages/service/package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "standard | snazzy && c8 --100 tap --no-coverage test/*test.js test/*/*test.mjs"
+    "test": "standard | snazzy && c8 --100 tap --no-coverage test/*test.js test/*/*.test.mjs"
   },
   "author": "Matteo Collina <hello@matteocollina.com>",
   "repository": {
