+++ packages/service/test/fixtures/missing-property.config.json
@@ -1,0 +1,1 @@
+{}
