+++ packages/utils/index.js
@@ -3,5 +3,6 @@
 const FileWatcher = require('./lib/file-watcher')
 const deepmerge = require('./lib/deepmerge')
 const isKeyEnabled = require('./lib/is-key-enabled')
+const createServerConfig = require('./lib/create-server-config')
 
-module.exports = { FileWatcher, deepmerge, isKeyEnabled }
+module.exports = { FileWatcher, deepmerge, isKeyEnabled, createServerConfig }
