+++ packages/utils/test/create-server-config.test.js
@@ -1,0 +1,12 @@
+'use strict'
+
+const { test } = require('tap')
+const { createServerConfig } = require('..')
+
+test('createServerConfig', async (t) => {
+  const data = createServerConfig({ server: { foo: 'bar' }, something: 'else' })
+  t.same(data, {
+    foo: 'bar',
+    something: 'else'
+  })
+})
